feat: Create dynamic questionnaire player
This change refactors the gun reform questionnaire into a generic, reusable component (`QuestionnaireScreen.js`) that can render any questionnaire structure passed to it as a navigation parameter.

The hardcoded data for the gun reform questionnaire has been moved to its own file (`data/gunReformNodes.js`).

This is the first phase in building a larger feature that will allow users to create and share their own questionnaires.
